Actualizacion070524 1230
Se actualiza la carpeta de pruebas, se avanza hasta el punto 09, no se prueba

+++ src/main/java/tads/Arista.java
@@ -1,17 +1,17 @@
 package tads;
 
 public class Arista {
-    public int kilometros;
+    public double kilometros;
 
-    public Arista(int kilometros) {
+    public Arista(double kilometros) {
         this.kilometros = kilometros;
     }
 
-    public int getKilometros() {
+    public double getKilometros() {
         return kilometros;
     }
 
-    public void setKilometros(int kilometros) {
+    public void setKilometros(double kilometros) {
         this.kilometros = kilometros;
     }
